Add `FeedImagesMapper` to Decouple Loader from API details
