+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -22,33 +22,14 @@ public final class RemoteFeedLoader: FeedLoader {
 		client.get(from: url) { result in
 			switch result {
 			case let .success((data, response)):
-				if response.statusCode == 200, let root = try? JSONDecoder().decode(Root.self, from: data) {
-					completion(.success(root.items.map { $0.item }))
+				if let images = try? FeedImagesMapper.map(data, response) {
+					completion(.success(images))
 				} else {
 					completion(.failure(Error.invalidData))
 				}
 			default:
 				completion(.failure(Error.connectivity))
 			}
-		}
-	}
-
-	private struct Root: Decodable {
-		let items: [Image]
-	}
-
-	private struct Image: Decodable {
-		let image_id: UUID
-		let image_desc: String?
-		let image_loc: String?
-		let image_url: URL
-
-		var item: FeedImage {
-			FeedImage(
-				id: image_id,
-				description: image_desc,
-				location: image_loc,
-				url: image_url)
 		}
 	}
 }
